self.countLabel.text gets string made from %lu rather than %d. Project settings updated from 5.0 to 5.1.

+++ Count Clicks/CCViewController.m
@@ -19,7 +19,7 @@
 - (IBAction)incrementClickCount:(id)sender
 {
     self.count ++;
-    self.countLabel.text = [NSString stringWithFormat:@"%d", self.count];
+    self.countLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)self.count];
 }
 
 - (void)viewDidLoad
